+++ test/scope_blue_print_test.dart
@@ -13,7 +13,7 @@ void main() {
       test('should provide a blue print with to nodes and one dependency', () {
         final scopeBluePrint = ScopeBluePrint.example();
         final dependency = scopeBluePrint.nodes.first;
-        final subScope = scopeBluePrint.childScopes.first;
+        final subScope = scopeBluePrint.children.first;
         final node = subScope.nodes.first as NodeBluePrint<int>;
         final customer = subScope.nodes.last as NodeBluePrint<int>;
         expect(scopeBluePrint.toString(), scopeBluePrint.key);
@@ -22,6 +22,120 @@ void main() {
         expect(customer.key, 'customer');
         expect(node.produce(<dynamic>[5], 0), 6);
         expect(customer.produce(<dynamic>[6], 0), 7);
+      });
+    });
+
+    group('fromJson', () {
+      test('should create a scope blue print from a JSON map', () {
+        final json = {
+          'a': {
+            'int': 5,
+            'double': 6.0,
+            'string': 'Hello',
+            'bool': true,
+            'bluePrint': const NodeBluePrint<int>(
+              key: 'bluePrint',
+              initialProduct: 8,
+            ),
+            'b': {
+              'c': {
+                'x': 123,
+              },
+            },
+            'c': const ScopeBluePrint(key: 'c'),
+          },
+        };
+
+        final scopeBluePrint = ScopeBluePrint.fromJson(json);
+        expect(scopeBluePrint.key, 'a');
+
+        void expectNode(int i, String key, dynamic value) {
+          expect(scopeBluePrint.nodes[i].key, key);
+          expect(scopeBluePrint.nodes[i].initialProduct, value);
+        }
+
+        expectNode(0, 'int', 5);
+        expectNode(1, 'double', 6.0);
+        expectNode(2, 'string', 'Hello');
+        expectNode(3, 'bool', true);
+        expectNode(4, 'bluePrint', 8);
+
+        expect(scopeBluePrint.children.length, 2);
+        expect(scopeBluePrint.children.first.key, 'b');
+        expect(scopeBluePrint.children.first.children.length, 1);
+        expect(scopeBluePrint.children.first.children.first.key, 'c');
+        expect(scopeBluePrint.children.last.key, 'c');
+
+        expect(
+          scopeBluePrint.children.first.children.first.nodes.first.key,
+          'x',
+        );
+
+        expect(
+          scopeBluePrint
+              .children.first.children.first.nodes.first.initialProduct,
+          123,
+        );
+      });
+
+      test('should assert that the key of a node and JSON are equal', () {
+        final json = {
+          'a': {
+            'int': const NodeBluePrint<int>(key: 'otherKey', initialProduct: 5),
+          },
+        };
+
+        expect(
+          () => ScopeBluePrint.fromJson(json),
+          throwsA(
+            isA<AssertionError>().having(
+              (e) => e.toString(),
+              'toString()',
+              contains('The key of the node "otherKey" must be "int".'),
+            ),
+          ),
+        );
+      });
+
+      test('should throw when in invalid type is provided', () {
+        final json = {
+          'a': {
+            'invalid': List<int>.empty(),
+          },
+        };
+
+        expect(
+          () => ScopeBluePrint.fromJson(json),
+          throwsA(
+            isA<ArgumentError>().having(
+              (e) => e.toString(),
+              'toString()',
+              contains(
+                'List<int>> not supported. '
+                'Use NodeBluePrint<_Map<String, List<int>>> instead.',
+              ),
+            ),
+          ),
+        );
+      });
+
+      test('should assert that the key of a scope and JSON are equal', () {
+        final json = {
+          'a': {
+            'key': const ScopeBluePrint(key: 'otherKey'),
+          },
+        };
+
+        expect(
+          () => ScopeBluePrint.fromJson(json),
+          throwsA(
+            isA<AssertionError>().having(
+              (e) => e.toString(),
+              'toString()',
+              contains('The key of the node "otherKey" must be "key".'),
+            ),
+          ),
+        );
       });
     });
 
@@ -84,13 +198,13 @@ void main() {
       });
 
       test('should return the node with the given key', () {
-        final bluePrint = ScopeBluePrint.example().childScopes.first;
+        final bluePrint = ScopeBluePrint.example().children.first;
         final node = bluePrint.findNode<int>('node');
         expect(node, isNotNull);
       });
 
       test('should throw if the type does not match', () {
-        final bluePrint = ScopeBluePrint.example().childScopes.first;
+        final bluePrint = ScopeBluePrint.example().children.first;
 
         expect(
           () => bluePrint.findNode<String>('node'),
@@ -123,11 +237,11 @@ void main() {
           final bluePrint = ScopeBluePrint.example();
           final otherSubScopes = <ScopeBluePrint>[];
           final copy = bluePrint.copyWith(subScopes: otherSubScopes);
-          expect(copy.childScopes, same(otherSubScopes));
+          expect(copy.children, same(otherSubScopes));
         });
 
         test('with the given overrides', () {
-          final bluePrint = ScopeBluePrint.example().childScopes.first;
+          final bluePrint = ScopeBluePrint.example().children.first;
           const overriddenNode = NodeBluePrint<int>(
             key: 'node',
             initialProduct: 5,